Adding support for `acl` plugin configuration
A plugin configuration can now be changed (it will be DELETEd and POSTed
again only if the configuration is changed with respect to what is
currently present.)

Each user can be added to one or more groups. When the user's key is
authenticated against an API, the groups he belongs to are added as an
header to send to the upstream server.